Control part of new api already kind of works
