+++ toxav/Makefile.inc
@@ -35,4 +35,21 @@ libtoxav_la_LIBADD =  libtoxcore.la \
                       $(PTHREAD_LIBS) \
                       $(AV_LIBS)
 
+
+noinst_PROGRAMS +=    av_test
+
+av_test_SOURCES =     ../toxav/av_test.c
+
+av_test_CFLAGS =      $(LIBSODIUM_CFLAGS) \
+                      $(NACL_CFLAGS)
+
+av_test_LDADD =       $(LIBSODIUM_LDFLAGS) \
+                      $(NACL_LDFLAGS) \
+                      libtoxav.la \
+                      libtoxcore.la \
+                      $(LIBSODIUM_LIBS) \
+                      $(NACL_OBJECTS) \
+                      $(NACL_LIBS)
+
+
 endif
